requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==2.2.6
django-braces==1.13.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
oauthlib==3.1.0
pathspec==0.6.0
Pillow==6.2.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.6
